+++ src/components/buttons/charactorBtn.tsx
